flashinfer a2a uses numel as trash index

+++ vllm/model_executor/layers/fused_moe/prepare_finalize/flashinfer_nvlink_two_sided.py
@@ -83,7 +83,9 @@ class FlashInferNVLinkTwoSidedPrepareAndFinalize(mk.FusedMoEPrepareAndFinalizeMo
         top_k = topk_ids.size(1)
 
         if lora_ids is not None:
-            topk_ids = topk_ids * self.lora_lanes + lora_ids.unsqueeze(1) + 1
+            topk_ids = (topk_ids * self.lora_lanes + lora_ids.unsqueeze(1) + 1).to(
+                topk_ids.dtype
+            )
             num_experts = num_experts * self.lora_lanes
 
         (self.alltoall_info, topk_ids, topk_weights, a1q, a1q_scale) = (
